refactor: modularise barbrief and update import paths in test files
This commit updates the import paths in test files to match the new module structure. The changes involve modifying patch decorators to use the correct import paths for document_reader, core, and brief_generator modules. The test_barbrief_command_parameters method was also updated to include the missing mock_read parameter.
