+++ litassist/commands/barbrief/__init__.py
@@ -1,0 +1,4 @@
+"""Barrister's brief generation command package."""
+from .core import barbrief
+
+__all__ = ["barbrief"]
